// src/app/models/log_inversa/bodegaVirtual.interface.ts
export interface BodegaVirtual {
    Ingreso_li:          string;
    Cliente:             string;
    Entrega:             string;
    Fecha_pedido:        Date;
    Fec_original_pedido: Date;
    Descripcion:         string;
    Cantidad:            number;
    Comuna:              string;
    Region:              string;
    Pistoleado:          boolean;
}